+++ plutarch-ledger-api/src/Plutarch/LedgerApi/V4.hs
@@ -7,8 +7,8 @@ prelude. These should be imported qualified.
 module Plutarch.LedgerApi.V4 (
   -- * Accounts
   Address.PAccountId (..),
-  PAccountBalanceInterval (..),
-  PAccountBalanceIntervals (..),
+  ContextsV4.PAccountBalanceInterval (..),
+  ContextsV4.PAccountBalanceIntervals (..),
 
   -- * Governance
   Contexts.PColdCommitteeCredential (..),
@@ -16,7 +16,7 @@ module Plutarch.LedgerApi.V4 (
   Contexts.PDRepCredential (..),
   Contexts.PDRep (..),
   Contexts.PDelegatee (..),
-  PTxCert (..),
+  ContextsV4.PTxCert (..),
   Contexts.PVoter (..),
   Contexts.PVote (..),
   Contexts.PGovernanceActionId (..),
@@ -28,11 +28,11 @@ module Plutarch.LedgerApi.V4 (
   Contexts.PProposalProcedure (..),
 
   -- * Context types
-  PScriptContext (..),
-  PScriptPurpose (..),
-  PScriptInfo (..),
-  PTopTxInfo (..),
-  PTopTxInfoSimplified (..),
+  ContextsV4.PScriptContext (..),
+  ContextsV4.PScriptPurpose (..),
+  ContextsV4.PScriptInfo (..),
+  ContextsV4.PTopTxInfo (..),
+  ContextsV4.PTopTxInfoSimplified (..),
 
   -- * Supporting types
 
@@ -58,10 +58,10 @@ module Plutarch.LedgerApi.V4 (
   Address.PAddress (..),
   Crypto.PPubKeyHash (..),
   V3Tx.PTxId (..),
-  PTxInfo (..),
+  ContextsV4.PTxInfo (..),
   Tx.PTxOut (..),
-  PTxOutRef (..),
-  PTxInInfo (..),
+  V3Tx.PTxOutRef (..),
+  ContextsV4.PTxInInfo (..),
   V2Tx.POutputDatum (..),
 
   -- ** Intervals
@@ -86,465 +86,18 @@ module Plutarch.LedgerApi.V4 (
   Scripts.PDatumHash (..),
 ) where
 
-import GHC.Generics (Generic)
-import Generics.SOP qualified as SOP
-import Plutarch.Internal.Lift (LiftError (CouldNotDecodeData))
-import Plutarch.LedgerApi.AssocMap (PUnsortedMap)
 import Plutarch.LedgerApi.AssocMap qualified as AssocMap
-import Plutarch.LedgerApi.Interval (PInterval)
 import Plutarch.LedgerApi.Interval qualified as Interval
-import Plutarch.LedgerApi.Utils (PMaybeData)
 import Plutarch.LedgerApi.Utils qualified as Utils
-import Plutarch.LedgerApi.V1.Credential (PCredential)
 import Plutarch.LedgerApi.V1.Credential qualified as Credential
-import Plutarch.LedgerApi.V1.Crypto (PPubKeyHash)
 import Plutarch.LedgerApi.V1.Crypto qualified as Crypto
-import Plutarch.LedgerApi.V1.Scripts (
-  PDatum,
-  PDatumHash,
-  PRedeemer,
-  PScriptHash,
- )
 import Plutarch.LedgerApi.V1.Scripts qualified as Scripts
-import Plutarch.LedgerApi.V1.Time (PPosixTime)
 import Plutarch.LedgerApi.V1.Time qualified as Time
 import Plutarch.LedgerApi.V2.Tx qualified as V2Tx
-import Plutarch.LedgerApi.V3.Contexts (
-  PColdCommitteeCredential,
-  PDRepCredential,
-  PDelegatee,
-  PGovernanceActionId,
-  PHotCommitteeCredential,
-  PProposalProcedure,
-  PVote,
-  PVoter,
- )
 import Plutarch.LedgerApi.V3.Contexts qualified as Contexts
-import Plutarch.LedgerApi.V3.MintValue (PMintValue)
 import Plutarch.LedgerApi.V3.MintValue qualified as MintValue
-import Plutarch.LedgerApi.V3.Tx (PTxId, PTxOutRef)
 import Plutarch.LedgerApi.V3.Tx qualified as V3Tx
-import Plutarch.LedgerApi.V4.Address (PAccountId)
 import Plutarch.LedgerApi.V4.Address qualified as Address
-import Plutarch.LedgerApi.V4.Tx (PTxOut)
+import Plutarch.LedgerApi.V4.Contexts qualified as ContextsV4
 import Plutarch.LedgerApi.V4.Tx qualified as Tx
-import Plutarch.LedgerApi.Value (PCurrencySymbol, PLovelace)
 import Plutarch.LedgerApi.Value qualified as Value
-import Plutarch.Prelude
-import PlutusLedgerApi.V4 qualified as Plutus
-
--- | @since 3.8.0
-data PAccountBalanceInterval (s :: S)
-  = PAccountBalanceLowerBound (Term s (PAsData PLovelace))
-  | PAccountBalanceUpperBound (Term s (PAsData PLovelace))
-  | PAccountBalanceBothBounds (Term s (PAsData PLovelace)) (Term s (PAsData PLovelace))
-  | PAccountBalanceExact (Term s (PAsData PLovelace))
-  deriving stock
-    ( -- | @since 3.8.0
-      Generic
-    )
-  deriving anyclass
-    ( -- | @since 3.8.0
-      SOP.Generic
-    , -- | @since 3.8.0
-      PIsData
-    , -- | @since 3.8.0
-      PEq
-    , -- | @since 3.8.0
-      PShow
-    )
-  deriving
-    ( -- | @since 3.8.0
-      PlutusType
-    , -- | @since 3.8.0
-      PValidateData
-    )
-    via (DeriveAsDataStruct PAccountBalanceInterval)
-
--- | @since 3.8.0
-deriving via
-  DeriveDataPLiftable PAccountBalanceInterval Plutus.AccountBalanceInterval
-  instance
-    PLiftable PAccountBalanceInterval
-
--- | @since 3.8.0
-instance PTryFrom PData (PAsData PAccountBalanceInterval)
-
--- | @since 3.8.0
-newtype PAccountBalanceIntervals (s :: S)
-  = PAccountBalanceIntervals (Term s (PUnsortedMap PAccountId PAccountBalanceInterval))
-  deriving stock
-    ( -- | @since 3.8.0
-      Generic
-    )
-  deriving anyclass
-    ( -- | @since 3.8.0
-      SOP.Generic
-    , -- | @since 3.8.0
-      PIsData
-    , -- | @since 3.8.0
-      PShow
-    )
-  deriving
-    ( -- | @since 3.8.0
-      PlutusType
-    )
-    via (DeriveNewtypePlutusType PAccountBalanceIntervals)
-  deriving
-    ( -- | @since 3.8.0
-      PValidateData
-    )
-    via (DeriveNewtypePValidateData PAccountBalanceIntervals (PUnsortedMap PAccountId PAccountBalanceInterval))
-
--- | @since 3.8.0
-instance PLiftable PAccountBalanceIntervals where
-  type AsHaskell PAccountBalanceIntervals = Plutus.AccountBalanceIntervals
-  type PlutusRepr PAccountBalanceIntervals = Plutus.Data
-  haskToRepr = Plutus.toData
-  reprToHask = maybe (Left CouldNotDecodeData) Right . Plutus.fromData
-  reprToPlut = reprToPlutUni
-  plutToRepr = plutToReprUni
-
--- | @since 3.8.0
-instance PTryFrom PData (PAsData PAccountBalanceIntervals)
-
--- | @since 3.8.0
-data PTxCert (s :: S)
-  = PTxCertRegAccount (Term s (PAsData PAccountId)) (Term s (PAsData PLovelace))
-  | PTxCertUnRegAccount (Term s (PAsData PAccountId)) (Term s (PAsData PLovelace))
-  | PTxCertDelegAccount (Term s (PAsData PAccountId)) (Term s (PAsData PDelegatee))
-  | PTxCertRegAccountDeleg (Term s (PAsData PAccountId)) (Term s (PAsData PDelegatee)) (Term s (PAsData PLovelace))
-  | PTxCertRegDRep (Term s (PAsData PDRepCredential)) (Term s (PAsData PLovelace))
-  | PTxCertUpdateDRep (Term s (PAsData PDRepCredential))
-  | PTxCertUnRegDRep (Term s (PAsData PDRepCredential)) (Term s (PAsData PLovelace))
-  | PTxCertPoolRegister (Term s (PAsData PPubKeyHash)) (Term s (PAsData PPubKeyHash))
-  | PTxCertPoolRetire (Term s (PAsData PPubKeyHash)) (Term s (PAsData PInteger))
-  | PTxCertAuthHotCommittee (Term s (PAsData PColdCommitteeCredential)) (Term s (PAsData PHotCommitteeCredential))
-  | PTxCertResignColdCommittee (Term s (PAsData PColdCommitteeCredential))
-  deriving stock
-    ( -- | @since 3.8.0
-      Generic
-    )
-  deriving anyclass
-    ( -- | @since 3.8.0
-      SOP.Generic
-    , -- | @since 3.8.0
-      PIsData
-    , -- | @since 3.8.0
-      PEq
-    , -- | @since 3.8.0
-      PShow
-    )
-  deriving
-    ( -- | @since 3.8.0
-      PlutusType
-    , -- | @since 3.8.0
-      PValidateData
-    )
-    via (DeriveAsDataStruct PTxCert)
-
--- | @since 3.8.0
-deriving via
-  DeriveDataPLiftable PTxCert Plutus.TxCert
-  instance
-    PLiftable PTxCert
-
--- | @since 3.8.0
-instance PTryFrom PData (PAsData PTxCert)
-
--- | @since 3.8.0
-data PScriptPurpose (s :: S)
-  = PMinting (Term s (PAsData PScriptHash)) (Term s (PAsData PCurrencySymbol))
-  | PSpending (Term s (PAsData PScriptHash)) (Term s (PAsData PTxOutRef))
-  | PWithdrawing (Term s (PAsData PScriptHash)) (Term s (PAsData PCredential))
-  | PCertifying (Term s (PAsData PScriptHash)) (Term s (PAsData PInteger)) (Term s (PAsData PTxCert))
-  | PVoting (Term s (PAsData PScriptHash)) (Term s (PAsData PVoter))
-  | PProposing (Term s (PAsData PScriptHash)) (Term s (PAsData PInteger)) (Term s (PAsData PProposalProcedure))
-  | PGuarding (Term s (PAsData PScriptHash)) (Term s (PAsData PInteger))
-  deriving stock
-    ( -- | @since 3.8.0
-      Generic
-    )
-  deriving anyclass
-    ( -- | @since 3.8.0
-      SOP.Generic
-    , -- | @since 3.8.0
-      PIsData
-    , -- | @since 3.8.0
-      PEq
-    , -- | @since 3.8.0
-      PShow
-    )
-  deriving
-    ( -- | @since 3.8.0
-      PlutusType
-    , -- | @since 3.8.0
-      PValidateData
-    )
-    via (DeriveAsDataStruct PScriptPurpose)
-
--- | @since 3.8.0
-deriving via
-  DeriveDataPLiftable PScriptPurpose Plutus.ScriptPurpose
-  instance
-    PLiftable PScriptPurpose
-
--- | @since 3.8.0
-instance PTryFrom PData (PAsData PScriptPurpose)
-
--- | @since 3.8.0
-data PTxInInfo (s :: S) = PTxInInfo
-  { ptxInInfo'outRef :: Term s (PAsData PTxOutRef)
-  , ptxInInfo'resolved :: Term s (PAsData PTxOut)
-  }
-  deriving stock
-    ( -- | @since 3.8.0
-      Generic
-    )
-  deriving anyclass
-    ( -- | @since 3.8.0
-      SOP.Generic
-    , -- | @since 3.8.0
-      PIsData
-    , -- | @since 3.8.0
-      PEq
-    , -- | @since 3.8.0
-      PShow
-    )
-  deriving
-    ( -- | @since 3.8.0
-      PlutusType
-    , -- | @since 3.8.0
-      PValidateData
-    )
-    via (DeriveAsDataStruct PTxInInfo)
-
--- | @since 3.8.0
-deriving via
-  DeriveDataPLiftable PTxInInfo Plutus.TxInInfo
-  instance
-    PLiftable PTxInInfo
-
--- | @since 3.8.0
-instance PTryFrom PData (PAsData PTxInInfo)
-
--- | @since 3.8.0
-data PTxInfo (s :: S) = PTxInfo
-  { ptxInfo'id :: Term s (PAsData PTxId)
-  , ptxInfo'subTxIx :: Term s (PAsData PInteger)
-  , ptxInfo'inputs :: Term s (PAsData (PBuiltinList (PAsData PTxInInfo)))
-  , ptxInfo'referenceInputs :: Term s (PAsData (PBuiltinList (PAsData PTxInInfo)))
-  , ptxInfo'outputs :: Term s (PAsData (PBuiltinList (PAsData PTxOut)))
-  , ptxInfo'fee :: Term s (PAsData PLovelace)
-  , ptxInfo'mint :: Term s (PAsData PMintValue)
-  , ptxInfo'txCerts :: Term s (PAsData (PBuiltinList (PAsData PTxCert)))
-  , ptxInfo'withdrawals :: Term s (PAsData (PUnsortedMap PAccountId PLovelace))
-  , ptxInfo'directDeposits :: Term s (PAsData (PUnsortedMap PAccountId PLovelace))
-  , ptxInfo'accountBalanceIntervals :: Term s (PAsData PAccountBalanceIntervals)
-  , ptxInfo'validRange :: Term s (PInterval PPosixTime)
-  , ptxInfo'guards :: Term s (PAsData (PBuiltinList (PAsData PCredential)))
-  , ptxInfo'requiredTopLevelGuards :: Term s (PAsData (PUnsortedMap PCredential (PMaybeData PDatum)))
-  , ptxInfo'redeemers :: Term s (PAsData (PUnsortedMap PScriptPurpose PRedeemer))
-  , ptxInfo'data :: Term s (PAsData (PUnsortedMap PDatumHash PDatum))
-  , ptxInfo'votes :: Term s (PAsData (PUnsortedMap PVoter (PUnsortedMap PGovernanceActionId PVote)))
-  , ptxInfo'proposalProcedures :: Term s (PAsData (PBuiltinList (PAsData PProposalProcedure)))
-  , ptxInfo'currentTreasuryAmount :: Term s (PMaybeData PLovelace)
-  , ptxInfo'treasuryDonation :: Term s (PAsData PLovelace)
-  }
-  deriving stock
-    ( -- | @since 3.8.0
-      Generic
-    )
-  deriving anyclass
-    ( -- | @since 3.8.0
-      SOP.Generic
-    , -- | @since 3.8.0
-      PIsData
-    , -- | @since 3.8.0
-      PEq
-    , -- | @since 3.8.0
-      PShow
-    )
-  deriving
-    ( -- | @since 3.8.0
-      PlutusType
-    , -- | @since 3.8.0
-      PValidateData
-    )
-    via (DeriveAsDataStruct PTxInfo)
-
--- | @since 3.8.0
-deriving via
-  DeriveDataPLiftable PTxInfo Plutus.TxInfo
-  instance
-    PLiftable PTxInfo
-
--- | @since 3.8.0
-instance PTryFrom PData (PAsData PTxInfo)
-
--- | @since 3.8.0
-data PTopTxInfoSimplified (s :: S) = PTopTxInfoSimplified
-  { pttis'ids :: Term s (PAsData (PBuiltinList (PAsData PTxId)))
-  , pttis'inputs :: Term s (PAsData (PBuiltinList (PAsData PTxInInfo)))
-  , pttis'referenceInputs :: Term s (PAsData (PBuiltinList (PAsData PTxInInfo)))
-  , pttis'outputs :: Term s (PAsData (PBuiltinList (PAsData PTxOut)))
-  , pttis'mints :: Term s (PAsData PMintValue)
-  , pttis'burns :: Term s (PAsData PMintValue)
-  , pttis'txCerts :: Term s (PAsData (PBuiltinList (PAsData PTxCert)))
-  , pttis'withdrawals :: Term s (PAsData (PUnsortedMap PAccountId PLovelace))
-  , pttis'directDeposits :: Term s (PAsData (PUnsortedMap PAccountId PLovelace))
-  , pttis'validRange :: Term s (PAsData (PInterval PPosixTime))
-  , pttis'guards :: Term s (PAsData (PUnsortedMap PCredential PUnit))
-  , pttis'scriptPurposes :: Term s (PAsData (PUnsortedMap PScriptPurpose PUnit))
-  , pttis'data :: Term s (PAsData (PUnsortedMap PDatumHash PDatum))
-  , pttis'votes :: Term s (PAsData (PUnsortedMap PVoter (PUnsortedMap PGovernanceActionId PVote)))
-  , pttis'proposalProcedures :: Term s (PAsData (PBuiltinList (PAsData PProposalProcedure)))
-  , pttis'currentTreasuryAmount :: Term s (PMaybeData PLovelace)
-  , pttis'treasuryDonations :: Term s (PAsData PLovelace)
-  }
-  deriving stock
-    ( -- | @since 3.8.0
-      Generic
-    )
-  deriving anyclass
-    ( -- | @since 3.8.0
-      SOP.Generic
-    , -- | @since 3.8.0
-      PIsData
-    , -- | @since 3.8.0
-      PEq
-    , -- | @since 3.8.0
-      PShow
-    )
-  deriving
-    ( -- | @since 3.8.0
-      PlutusType
-    , -- | @since 3.8.0
-      PValidateData
-    )
-    via (DeriveAsDataStruct PTopTxInfoSimplified)
-
--- | @since 3.8.0
-deriving via
-  DeriveDataPLiftable PTopTxInfoSimplified Plutus.TopTxInfoSimplified
-  instance
-    PLiftable PTopTxInfoSimplified
-
--- | @since 3.8.0
-instance PTryFrom PData (PAsData PTopTxInfoSimplified)
-
--- | @since 3.8.0
-data PTopTxInfo (s :: S) = PTopTxInfo
-  { topTxInfo'subTransactions :: Term s (PAsData (PBuiltinList (PAsData PTxInfo)))
-  , topTxInfo'datums :: Term s (PAsData (PUnsortedMap PInteger PDatum))
-  , topTxInfo'startingBalanceIntervals :: Term s (PAsData PAccountBalanceIntervals)
-  , topTxInfo'simplified :: Term s (PAsData PTopTxInfoSimplified)
-  }
-  deriving stock
-    ( -- | @since 3.8.0
-      Generic
-    )
-  deriving anyclass
-    ( -- | @since 3.8.0
-      SOP.Generic
-    , -- | @since 3.8.0
-      PIsData
-    , -- | @since 3.8.0
-      PEq
-    , -- | @since 3.8.0
-      PShow
-    )
-  deriving
-    ( -- | @since 3.8.0
-      PlutusType
-    , -- | @since 3.8.0
-      PValidateData
-    )
-    via (DeriveAsDataStruct PTopTxInfo)
-
--- | @since 3.8.0
-deriving via
-  DeriveDataPLiftable PTopTxInfo Plutus.TopTxInfo
-  instance
-    PLiftable PTopTxInfo
-
--- | @since 3.8.0
-instance PTryFrom PData (PAsData PTopTxInfo)
-
--- | @since 3.8.0
-data PScriptInfo (s :: S)
-  = PMintingScript (Term s (PAsData PCurrencySymbol))
-  | PSpendingScript (Term s (PAsData PTxOutRef)) (Term s (PMaybeData PDatum))
-  | PWithdrawingScript (Term s (PAsData PAccountId))
-  | PCertifyingScript (Term s (PAsData PInteger)) (Term s (PAsData PTxCert))
-  | PVotingScript (Term s (PAsData PVoter))
-  | PProposingScript (Term s (PAsData PInteger)) (Term s (PAsData PProposalProcedure))
-  | PGuardingScript (Term s (PAsData PInteger)) (Term s (PMaybeData PTopTxInfo))
-  deriving stock
-    ( -- | @since 3.8.0
-      Generic
-    )
-  deriving anyclass
-    ( -- | @since 3.8.0
-      SOP.Generic
-    , -- | @since 3.8.0
-      PIsData
-    , -- | @since 3.8.0
-      PEq
-    , -- | @since 3.8.0
-      PShow
-    )
-  deriving
-    ( -- | @since 3.8.0
-      PlutusType
-    , -- | @since 3.8.0
-      PValidateData
-    )
-    via (DeriveAsDataStruct PScriptInfo)
-
--- | @since 3.8.0
-deriving via
-  DeriveDataPLiftable PScriptInfo Plutus.ScriptInfo
-  instance
-    PLiftable PScriptInfo
-
--- | @since 3.8.0
-instance PTryFrom PData (PAsData PScriptInfo)
-
--- | @since 3.8.0
-data PScriptContext (s :: S) = PScriptContext
-  { pscriptContext'txInfo :: Term s (PAsData PTxInfo)
-  , pscriptContext'redeemer :: Term s (PAsData PRedeemer)
-  , pscriptContext'scriptInfo :: Term s (PAsData PScriptInfo)
-  , pscriptContext'scriptHash :: Term s (PAsData PScriptHash)
-  }
-  deriving stock
-    ( -- | @since 3.8.0
-      Generic
-    )
-  deriving anyclass
-    ( -- | @since 3.8.0
-      SOP.Generic
-    , -- | @since 3.8.0
-      PIsData
-    , -- | @since 3.8.0
-      PEq
-    , -- | @since 3.8.0
-      PShow
-    )
-  deriving
-    ( -- | @since 3.8.0
-      PlutusType
-    , -- | @since 3.8.0
-      PValidateData
-    )
-    via (DeriveAsDataStruct PScriptContext)
-
--- | @since 3.8.0
-deriving via
-  DeriveDataPLiftable PScriptContext Plutus.ScriptContext
-  instance
-    PLiftable PScriptContext
-
--- | @since 3.8.0
-instance PTryFrom PData (PAsData PScriptContext)
